donations_platform/requirements.txt
Django==4.2
djangorestframework==3.14.0
python-decouple==3.5
django-environ==0.8.1
django-extensions==3.2.1
Pillow==10.0.0
django-allauth==0.52.0
```